+++ src/EntityLocator.php
@@ -5,10 +5,12 @@ namespace Somnambulist\ApiClient;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerAwareTrait;
 use Psr\Log\LogLevel;
-use Somnambulist\ApiClient\Client\ApiRequestHelper;
 use Somnambulist\ApiClient\Behaviours\EntityLocator\CanAppendIncludes;
 use Somnambulist\ApiClient\Behaviours\EntityLocator\HydrateAsCollection;
 use Somnambulist\ApiClient\Behaviours\EntityLocator\HydrateSingleObject;
+use Somnambulist\ApiClient\Behaviours\LoggerWrapper;
+use Somnambulist\ApiClient\Behaviours\RoutePrefixer;
+use Somnambulist\ApiClient\Client\ApiRequestHelper;
 use Somnambulist\ApiClient\Contracts\ApiClientInterface;
 use Somnambulist\ApiClient\Contracts\EntityLocatorInterface;
 use Somnambulist\ApiClient\Mapper\ObjectMapper;
@@ -16,7 +18,6 @@ use Somnambulist\Collection\Contracts\Collection;
 use Somnambulist\Collection\MutableCollection;
 use Symfony\Component\HttpClient\Exception\ClientException;
 use function array_merge;
-use function sprintf;
 
 /**
  * Class EntityLocator
@@ -32,7 +33,7 @@ use function sprintf;
  * be requested from the API end point. Note that this requires support from the Api
  * end point.
  *
- * @package Somnambulist\ApiClient\Client
+ * @package    Somnambulist\ApiClient\Client
  * @subpackage Somnambulist\ApiClient\Client\EntityLocator
  */
 class EntityLocator implements LoggerAwareInterface, EntityLocatorInterface
@@ -42,6 +43,8 @@ class EntityLocator implements LoggerAwareInterface, EntityLocatorInterface
     use HydrateAsCollection;
     use HydrateSingleObject;
     use LoggerAwareTrait;
+    use LoggerWrapper;
+    use RoutePrefixer;
 
     /**
      * @var ApiClientInterface
@@ -59,14 +62,18 @@ class EntityLocator implements LoggerAwareInterface, EntityLocatorInterface
     protected $apiHelper;
 
     /**
+     * The type of objects that will be hydrated
+     *
      * @var string
      */
     protected $className;
 
     /**
+     * The name of the field / key for the primary id of the mapped objects
+     *
      * @var string
      */
-    protected $classPrimaryKey;
+    protected $identityField;
 
     /**
      * The collection class to return when hydrating collections
@@ -75,20 +82,12 @@ class EntityLocator implements LoggerAwareInterface, EntityLocatorInterface
      */
     protected $collectionClass = MutableCollection::class;
 
-    /**
-     * Constructor
-     *
-     * @param ApiClientInterface $client
-     * @param ObjectMapper       $mapper
-     * @param string             $class
-     * @param string             $identity
-     */
     public function __construct(ApiClientInterface $client, ObjectMapper $mapper, string $class, string $identity = 'id')
     {
-        $this->mapper          = $mapper;
-        $this->client          = $client;
-        $this->className       = $class;
-        $this->classPrimaryKey = $identity;
+        $this->mapper        = $mapper;
+        $this->client        = $client;
+        $this->className     = $class;
+        $this->identityField = $identity;
 
         $this->apiHelper = new ApiRequestHelper();
     }
@@ -107,7 +106,7 @@ class EntityLocator implements LoggerAwareInterface, EntityLocatorInterface
 
     public function find($id): ?object
     {
-        $options = [$this->classPrimaryKey => (string)$id];
+        $options = [$this->identityField => (string)$id];
 
         try {
             $response = $this->client->get($this->prefix('view'), $this->appendIncludes($options));
@@ -115,8 +114,8 @@ class EntityLocator implements LoggerAwareInterface, EntityLocatorInterface
             return $this->hydrateObject($response);
         } catch (ClientException $e) {
             $this->log(LogLevel::ERROR, $e->getMessage(), [
-                'route'                => $this->client->route($this->prefix('view'), $this->appendIncludes($options)),
-                $this->classPrimaryKey => (string)$id,
+                'route'              => $this->client->route($this->prefix('view'), $this->appendIncludes($options)),
+                $this->identityField => (string)$id,
             ]);
         }
 
@@ -147,19 +146,5 @@ class EntityLocator implements LoggerAwareInterface, EntityLocatorInterface
     public function findOneBy(array $criteria, array $orderBy = []): ?object
     {
         return $this->findBy($criteria, $orderBy, 1)->first() ?: null;
-    }
-
-
-
-    protected function prefix(string $route): string
-    {
-        return sprintf('%s.%s', $this->client->router()->service()->alias(), $route);
-    }
-
-    protected function log(string $level, string $message, array $context = []): void
-    {
-        if ($this->logger) {
-            $this->logger->log($level, $message, $context);
-        }
     }
 }